LPS-133275 Move to correct package
